+++ docs/changelog.md
@@ -20,6 +20,7 @@ brief notes about them may accumulate here.
 * Wayland: fixed window decoration and full screen handling [#224](https://github.com/wez/wezterm/issues/224)
 * Wayland: fixed an issue where key repeat processing could "run away" and hang the application
 * New: added [foreground_text_hsb](config/lua/config/foreground_text_hsb.md) setting to adjust hue, saturation and brightness when text is rendered.
+* New: added [ResetFontAndWindowSize](config/lua/keyassignment/ResetFontAndWindowSize.md) key assignment.
 
 ### 20210203-095643-70a364eb